Started emulating the CPU (#4)
This fully emulates the instruction set of the CPU at the heart of the NES. It emulates instructions one at a time. It passes the blargg unit tests for instructions as well as most of the nestest, but there's incomplete support for unofficial opcodes.

This branch does a LOT of things, since as I'm writing the CPU core I'm also figuring out a build system, a way to keep my code clean and trying to understand how to extract more performance out of Python. After all, the ultimate goal is to try to get the emulator to be usable with CPython, not just pypy.

The code quality is taken care of by the pre-commit hooks that I've installed. This means flake8, black code style, pycodestyle and a few other pre-commit hooks.

To understand the architecture of the emulator and it's modules and submodules, see the files under the docs folder. If you start reviewing the code by looking at the doc, you'll get a much better understanding of how the code is structured.